+++ header.h
@@ -17,27 +17,27 @@ public:
     void modificacion();
     void setModificado(bool modificado);
     bool getModificado() const;
-    void setSize(int size);
-    int getSize() const;
+    void setSize(short size);
+    unsigned short getSize() const;
     void setNumregtot(int numregtot);
-    int getNumregtot() const;
+    unsigned int getNumregtot() const;
     void setNumreglibres(int numreglibres);
-    int getNumreglibres() const;
+    unsigned int getNumreglibres() const;
     void setRegistros_libres(int registros_libres);
-    int getRegistros_libres() const;
+    unsigned int getRegistros_libres() const;
     void setFin(int fin);
-    int getFin() const;
+    unsigned int getFin() const;
     void setInicio(int inicio);
-    int getInicio() const; //Devuelve el contenido de una carpeta
+    unsigned int getInicio() const; //Devuelve el contenido de una carpeta
         
 private:
     private:
-        int inicio; 
-        int fin; 
-        int registros_libres;
-        int numreglibres;
-        int numregtot;
-        int size;
+        unsigned int inicio; 
+        unsigned int fin; 
+        unsigned int registros_libres;
+        unsigned int numreglibres;
+        unsigned int numregtot;
+        unsigned short size;
         bool modificado= false;
 
 };
